Update game
Henry clarified that all the def methods that are defined in Leo's menu need to go ABOVE the def method for the menu.  Otherwise, the menu can't "talk to" defined methods that are below it in the code.

Henry also said I needed to have a separate Player.rb file and then do 'require relative player' at the top of the main Casino game.rb, so I did that. 

He also said the whole game should be contained within the class of Casino, with an 'end' at the very bottom, so I did that.

----- STILL NOT RUNNING CORRECTLY -----  as of 11:30pm EST on Wed night before this project is DUE TOMORROW :(